requirements.txt
blinker==1.9.0
click==8.3.1
Flask==3.1.2
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.3
packaging==25.0
pytz==2025.2
RPi.bme280==0.2.4
smbus2==0.6.0
Werkzeug==3.1.4
